refactor: Copy DOMParserOptions `locator` and `xmlns` to instance
Instead of accessing `this.options` in `DOMParser.parseToString`,
the default values are now applied in the constructor.
Since the locator passed to `options` is no longer being modified,
the type of the option was changed to boolean.
There is no change in behavior in this commit,
since truthy and falsy values are accepted as well.