refactor: Drop `DOMParser.options` property
All options are now taken care of by the constructor and are available as individual properties.
Most are marked as `readonly`, some are `private`.

BREAKING CHANGE: If you used `DOMParser.options` after creating an instance.
You can still read the individual properties from the instance,
but there is no longer a way to mutate them, so you need to really pass the required options when constructing them.